+++ lib/schleuder/list.rb
@@ -291,7 +291,7 @@ module Schleuder
         Schleuder.logger.info "Deleted listdir"
       else
         # Don't use list-logger here — if the list-dir isn't present we can't log to it!
-        Schleuder.logger.info "Couldn't delete listdir, directly not present"
+        Schleuder.logger.info "Couldn't delete listdir, directory not present"
       end
       true
     rescue => exc